Updated Header & Project Layout
I chose to make card groups of 2 instead of 3, so as I add new projects, there's less empty space. I am debating changing the order of my projects from "oldest to newest" to "newest to oldest" as well. I think having the most recent projects on the top may be more effective for the person reviewing my portfolio. When I go to do that transition, it will be best if I save the code for each card individually as well, so future updates do not get messed up.