Updated the Projects order to Newest to Oldest
I renamed a photo, updated the order and added the order to my header to clarify for potential reviewers. 

I made this project public to help others trying to navigate making their own portfolio from scratch. From this point on, I will no longer be committing changes made to my portfolio. Feel free to reach out to me here if you have questions. My goal was to help others, showcase a project that was not school-related, and that I am able to make updated changes and label the commits properly. 

Best wishes to all whom use this project as a detailed template on your journeys. 
~Many missteps were made, corrected and at the end of the day, my efforts are something I am proud of!